git-svn-id: https://svn.imp.fu-berlin.de/cmb/stallone@764 3c54fc69-7031-4ae6-9a21-c9b007ec3f0a

+++ src/stallone/ints/PrimitiveIntTools.java
@@ -1261,7 +1261,7 @@ public class PrimitiveIntTools
     {
         if (start > end)
             throw(new IllegalArgumentException("Invalid range: ["+start+", "+end+", "+step+"]"));
-        int[] res = new int[(int) Math.ceil((end - start) / step)];
+        int[] res = new int[(int) Math.ceil((double)(end - start) / (double)step)];
 
         if (res.length > 0)
         {
